requirements.txt
evdev
PyYaml
requests

